Refactor lsp code into its own module
